add spatie role and permissions

+++ routes/dashboard.php
@@ -83,25 +83,29 @@ Route::middleware('auth')->prefix('dashboard')->group(function ()
 	 */
 	Route::prefix('user')->as('management.user.')->group(function() {
 		// Route::post('/user/{user_id}/follow', [UsersController::class, 'followUser']);
-		Route::put('/update-socials', [UsersController::class, 'update_socials'])->name('update_socials');
-		Route::get('/profile', [UsersController::class, 'profile'])->name('profile');
-		Route::get('/settings', [UsersController::class, 'settings'])->name('settings');
-		Route::get('/profile/{user}', [UsersController::class, "show"])->name('show');
-
+		Route::middleware("permission:user_show")->group(function() {
+			Route::get('/profile', [UsersController::class, 'profile'])->name('profile');
+			Route::get('/settings', [UsersController::class, 'settings'])->name('settings');
+			Route::get('/profile/{user}', [UsersController::class, "show"])->name('show');
+		});
+		
 		// CRUD
-		Route::post('/change-password', [UsersController::class, 'change_password'])->name('change_pass');
 		// Can update
-		Route::get('/{user}/edit', [UsersController::class, 'edit_user'])->middleware('can:update,App\Models\User,user');
-		Route::post('/{user}/update', [UsersController::class, 'update'])->middleware('can:update,App\Models\User,user');
+		Route::middleware("permission:user_edit")->group(function() {
+			Route::put('/update-socials', [UsersController::class, 'update_socials'])->name('update_socials');
+			Route::post('/change-password', [UsersController::class, 'change_password'])->name('change_pass');
+			Route::get('/{user}/edit', [UsersController::class, 'edit_user']);
+			Route::post('/{user}/update', [UsersController::class, 'update']);
+		});
 		// Can delete
-		Route::post('/delete', [UsersController::class, 'delete'])->middleware('can:delete,App\Models\User');
+		Route::post('/delete', [UsersController::class, 'delete'])->middleware('permission:user_delete');
 		// Can view
-		Route::middleware("can:view,App\Models\User")->group(function() {
+		Route::middleware("permission:user_access")->group(function() {
 			Route::get('/cards', [UsersController::class, 'cards'])->name('cards');
 			Route::get('/list', [UsersController::class, 'index'])->name('list');
 		});
 		// Can create
-		Route::middleware("can:create,App\Models\User")->group(function() {
+		Route::middleware("permission:user_create")->group(function() {
 			Route::get('/new', [UsersController::class, 'create'])->name('new');
 			Route::post('/new', [UsersController::class, 'store'])->name('store');
 		});
@@ -109,8 +113,8 @@ Route::middleware('auth')->prefix('dashboard')->group(function ()
 
 
 	Route::prefix('image')->as('image.')->group(function() {
-		Route::post('/new/slider', [ImagesController::class, "storeSlider"])->name("storeSlider");
-		Route::delete('/delete/{image}', [ImagesController::class, "destroy"])->name("destroy");
+		Route::post('/new/slider', [ImagesController::class, "storeSlider"])->name("storeSlider")->middleware( "permission:image_upload_slider");
+		Route::delete('/delete/{image}', [ImagesController::class, "destroy"])->name("destroy")->middleware("permission:image_upload_delete");
 	});
